Add --open-playlist-string option
Lets playlists be passed to `play` by an actual string argument; e.g.
--playlist-string '{"tracks": [...]}'. Could work well with http-music's
own crawler utilities (e.g. --playlist-string
'{"source": ["crawl-local", "."]}' or with external ones (e.g.
--playlist-string "$(cool-crawler-utilith.sh)".

At the moment the only workaround is to use a temporary file to store the
playlist in, or to pipe the output of the crawler command (which might be
`echo`!) to http-music and use /dev/stdin as the --playlist-file value..
which obviously isn't cross-platform or practical!

I'd like to have some aliases in the future since
--open-playlist-string isn't really easy to type.